Add files via upload
The Flutar.com Project was an initiative by the Rocket team to consolidate and practice fundamental concepts of web development, focusing on the Box Model and visual design. We explored details such as borders, positioning, colors, and fonts, enhancing our understanding of how to control page layout. Additionally, we successfully included a favicon, marking an important milestone in our learning journey. This project reflects our commitment to excellence and continuous growth in web development.